requirements.txt
python-binance
PyYAML
pandas
